Fixing boundary errors
When working with character positions in rules, like xNM, iNX, ZN etc., hashcat uses A-Z as placeholders for numbers from 10 to 34. In the hcr code, that was not taken into consideration. A lot of the positions were handled with 

pos = int(rules[idx])

which uses base 10 and gives issues with positions over 9. Other places this was handled with 

pos = int(rules[idx],16)

which uses base 16 and gives issues on positions over 15. Those have been replaced with

alphanum_list.index(rules[idx].upper())

The variable alphanum_list is declared in the top of the transform function like

alphanum_list = list('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ')

The .N rule (Replace character @ N with value at @ N plus 1) gave an error if the character at position N was the last character in the string, because then it would look outside of the list giving an overflow error. This was fixed by ignoring the rule if boundaries were breached.